DHStore is another valuestore implementation
DHStore is no longer an engine option, but is now another valuestore implementation.

Fixes #169
