+++ store/dhstore/client/http_req.go
@@ -1,11 +1,11 @@
-// Package dhstore forks the HTTP request models of dhtsotre in order to
+// Package client forks the HTTP request models of dhtsotre in order to
 // avoid forcing upstream projects into C-bindings required by Foundation DB.
 // Note that this repo only needs the Write request models as it only ever
 // persists data to dhstore. It does not look up records from it.
 //
 // The structs here are compatible with the API specification of dhstore.
 // See: https://github.com/ipni/dhstore/blob/main/openapi.yaml
-package dhstore
+package client
 
 import (
 	"github.com/multiformats/go-multihash"
